% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloomjoin.R
\name{bloom_join}
\alias{bloom_join}
\title{Perform a join between two data frames using Bloom filter optimization}
\usage{
bloom_join(
  x,
  y,
  by = NULL,
  type = "inner",
  bloom_size = NULL,
  false_positive_rate = 0.01,
  verbose = TRUE
)
}
\arguments{
\item{x}{The larger data frame}

\item{y}{The smaller data frame}

\item{by}{Character vector of variables to join by}

\item{type}{The type of join: "inner", "left", "right", or "full"}

\item{bloom_size}{Bloom filter size, defaults to size of y}

\item{false_positive_rate}{Acceptable rate of false positives (between 0 and 1)}

\item{verbose}{Whether to print information about the filtering process}
}
\value{
A data frame
}
\description{
Perform a join between two data frames using Bloom filter optimization
}
\examples{
# Create test data
large_df <- data.frame(id = 1:1000, value = runif(1000))
small_df <- data.frame(id = sample(1:2000, 100), info = letters[1:100])

# Perform bloom filter join
result <- bloom_join(large_df, small_df, by = "id")
}
