% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloomjoin.R
\docType{package}
\name{bloomjoin-package}
\alias{bloomjoin}
\alias{bloomjoin-package}
\title{bloomjoin: Efficient data.frame joins using Bloom filters}
\description{
This package provides optimized join operations for large data frames using
Bloom filters to reduce memory usage and improve performance.
}
\author{
\strong{Maintainer}: Gaurav Sood \email{gsood07@gmail.com}

}
\keyword{internal}
