% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloomjoin.R
\name{find_closest_match}
\alias{find_closest_match}
\title{Find closest match for error suggestions}
\usage{
find_closest_match(input, options)
}
\description{
Find closest match for error suggestions
}
\keyword{internal}
