% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloomjoin.R
\name{perform_standard_join}
\alias{perform_standard_join}
\title{Perform standard join operations}
\usage{
perform_standard_join(x, y, by = NULL, type = "inner")
}
\arguments{
\item{x}{A data frame}

\item{y}{A data frame}

\item{by}{A character vector of variables to join by}

\item{type}{Type of join: "inner", "left", "right", "full", "semi", or "anti"}
}
\value{
A joined data frame
}
\description{
Internal function to perform standard join operations using dplyr
}
\keyword{internal}
