% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloomjoin.R
\name{prepare_join_columns}
\alias{prepare_join_columns}
\title{Prepare join columns and handle named vectors}
\usage{
prepare_join_columns(x, y, by)
}
\arguments{
\item{x}{A data frame}

\item{y}{A data frame}

\item{by}{A character vector of variables to join by}
}
\value{
List with prepared join columns
}
\description{
Prepare join columns and handle named vectors
}
\keyword{internal}
