% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloomjoin.R
\name{should_use_bloom_filter}
\alias{should_use_bloom_filter}
\title{Determine if Bloom filter should be used for the join}
\usage{
should_use_bloom_filter(x, y, type)
}
\arguments{
\item{x}{A data frame}

\item{y}{A data frame}

\item{type}{Type of join}
}
\value{
Logical indicating whether to use Bloom filter
}
\description{
Internal function to decide whether a Bloom filter will be beneficial
}
\keyword{internal}
