% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloomjoin.R
\name{validate_and_prepare_join}
\alias{validate_and_prepare_join}
\title{Validate inputs and prepare join parameters}
\usage{
validate_and_prepare_join(x, y, by, type, false_positive_rate, verbose)
}
\arguments{
\item{x}{A data frame}

\item{y}{A data frame}

\item{by}{A character vector of variables to join by}

\item{type}{Type of join}

\item{false_positive_rate}{False positive rate}

\item{verbose}{Logical, whether to print performance information}
}
\value{
List with validation results and prepared parameters
}
\description{
Validate inputs and prepare join parameters
}
\keyword{internal}
