% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloomjoin.R
\name{validate_data_frames}
\alias{validate_data_frames}
\title{Validate that inputs are data frames}
\usage{
validate_data_frames(x, y)
}
\description{
Validate that inputs are data frames
}
\keyword{internal}
