% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloomjoin.R
\name{validate_join_parameters}
\alias{validate_join_parameters}
\title{Validate join parameters}
\usage{
validate_join_parameters(type, false_positive_rate)
}
\description{
Validate join parameters
}
\keyword{internal}
