% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloomjoin.R
\name{bloom_join}
\alias{bloom_join}
\title{Join two data frames using a hash-based filtering approach similar to Bloom filters}
\usage{
bloom_join(
  x,
  y,
  by = NULL,
  type = "inner",
  bloom_size = NULL,
  false_positive_rate = 0.01,
  verbose = TRUE
)
}
\arguments{
\item{x}{A data frame}

\item{y}{A data frame}

\item{by}{A character vector of variables to join by}

\item{type}{Type of join: "inner", "left", "right", "full", "semi", or "anti"}

\item{bloom_size}{Size of the Bloom filter. Defaults to the number of rows in y}

\item{false_positive_rate}{False positive rate for the Bloom filter}

\item{verbose}{Whether to print progress messages}
}
\value{
A joined data frame with a class of "bloomjoin" and additional attributes
}
\description{
Join two data frames using a hash-based filtering approach similar to Bloom filters
}
\examples{
x <- data.frame(id = 1:100, value_x = rnorm(100))
y <- data.frame(id = 50:150, value_y = rnorm(101))
result <- bloom_join(x, y, by = "id")
}
