% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloomjoin.R
\name{bloom_join}
\alias{bloom_join}
\title{Join two data frames using a Bloom filter implemented with Rcpp}
\usage{
bloom_join(
  x,
  y,
  by = NULL,
  type = "inner",
  bloom_size = NULL,
  false_positive_rate = 0.01,
  verbose = FALSE
)
}
\arguments{
\item{x}{A data frame}

\item{y}{A data frame}

\item{by}{A character vector of variables to join by}

\item{type}{Type of join: "inner", "left", "right", "full", "semi", or "anti"}

\item{bloom_size}{Size of the Bloom filter. If NULL, will be automatically calculated}

\item{false_positive_rate}{False positive rate for the Bloom filter (default: 0.01)}

\item{verbose}{Logical, whether to print performance information (default: FALSE)}
}
\value{
A joined data frame with a class of "bloomjoin"
}
\description{
This function uses a high-performance Rcpp implementation of Bloom filters to efficiently join
large data frames. The Bloom filter is used to pre-filter the left data frame, reducing the
number of comparisons needed for the actual join.
}
\examples{
x <- data.frame(id = 1:100000, value_x = rnorm(100000))
y <- data.frame(id = 50001:60000, value_y = rnorm(10000))
result <- bloom_join(x, y, by = "id")
}
