% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloomjoin.R
\name{bloom_params}
\alias{bloom_params}
\title{Compute Bloom filter parameters}
\usage{
 bloom_params(n, p = 0.01)
}
\arguments{
\item{n}{Estimated number of elements that will be inserted into the filter.}

\item{p}{Desired false positive rate. Defaults to 0.01.}
}
\value{
A list with the number of bits (\code{m}) and hash functions (\code{k}).
}
\description{
Return the canonical Bloom filter parameters for a target number of elements and
false positive rate.
}
\examples{
bloom_params(1e6, p = 0.001)
}
